Dialogue Box Close + Preview Setup
Dialogue box will now close after any action is set up, in addition there is now a list that shows red tiles based on the number of items in the image list. This is to be updated soon with a list of the actual images loaded into memory. One thing to look out for later is that the list doesn't get cleared when we leave the screen currently. Need to clear it after submit or when we leave the screen; however, don't clear it when the user is in their camera/gallery sections. Also removed the assist dart file and moved it into report screen due to me realizing no other file will use this code and it makes it a hassle to pass information around.