Submit exits report screen
It will reset the data in the screen now so if the report screen is opened at later date it will have a fresh report screen to work with. A little bit of testing and talking with Keshaun revealed that the list does update actively on an android device, but not emulator. Its likely that the reason this occuring is because imageFileList is being updated by async function and the widget is upating, but its not seeing data before it updates. This can be seen by the new setstate code being called on button press.