Report Reminders Local Notification
So there is a templated local notification that now occurs when you select a time reminder for later. Do note currently that it displays template text, only works if the application is running or open in the background, and does not default on whether or not to show a banner at the top of the android device (need to figure out the permissions properly for this still). So it 'works' and I'm likely about to head off soon, so wanted to push what I have so far.