Reminder to Report Later Done
Implemented to create an alert that will open the app (if enabled). When an alert is created it will also create a notification that is added to the notification list at a later time. (This presumes the app is running, it will work while in background, but turning off app it will reset to default. Which is fine for demo purposes). Current known issue with the alert that appears where it can only schedule one, but the notifications appearing on the notification screen do not have this issue.