Changed Report Display
Title is based on report type now, also have it display a readable version. There is probably a more efficient way to do this because at the moment if we want to add a new report type to the list we have to not only go to report submission and fix up that list, but we need to do the same here now. In addition don't really know if the way I'm doing it is particular... efficient. I noticed that I can change the icon as well this way, but with how I originally coded it, it was causing them to flash between the original icon and the one that was being set. Might be able to talk to Keshaun about this one.