Psuedo Implemented Heatmaps
It makes a heatmap, albeit not based on the spots I was intending. There is difficulty extracting the reports by location / association id and then using the reports' information to create points. May need Keshaun to look at this or give up on the idea, but right now brain needs sleep. Right now it will just draw a line of points based for each "location" in the database. Also currently have ! check in the analysis_screen removed so don't have to keep logging in anytime the app needs to restart while testing.