Implemented Heatmaps
A heatmap has been implemented to set heatmap points at the location of reports that have been tagged with a location id. Those points are weighted with the intensity based on the location they belong to. Currently only one location has an intensity of not 0, so need to find a way to make sure those are being updated.